requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
boto3==1.35.63
botocore==1.35.63
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.5.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
dataclasses-json==0.6.7
distlib==0.3.9
distro==1.9.0
dulwich==0.21.7
fastapi==0.115.5
fastjsonschema==2.20.0
ffmpeg==1.4
filelock==3.16.1
frozenlist==1.5.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
Jinja2==3.1.5
jiter==0.7.1
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
keyring==24.3.1
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langsmith==0.1.143
MarkupSafe==3.0.2
marshmallow==3.23.1
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.54.4
orjson==3.10.11
packaging==24.2
pexpect==4.9.0
pkginfo==1.11.2
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
propcache==0.2.0
proto-plus==1.26.0
protobuf==5.29.3
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyparsing==3.2.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.10.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
SQLAlchemy==2.0.35
starlette==0.41.2
tenacity==9.0.0
tiktoken==0.8.0
tomlkit==0.13.2
tqdm==4.67.0
trove-classifiers==2024.10.21.16
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
virtualenv==20.27.1
xattr==1.1.0
yarl==1.17.1
